Create registration & login page

+++ backend/routes/trainings.js
@@ -1,5 +1,10 @@
 import express from "express";
 import Training from "../models/training.js";
+import {
+  isTrainingValid,
+  isStartAndEndDateValid,
+  isStartAndEndTimeValid,
+} from "../controllers/trainings.js";
 
 const router = express.Router();
 
@@ -12,9 +17,13 @@ router.get("/", async (req, res) => {
   }
 });
 
-router.get("/my-trainings", async (req, res) => {
+router.get("/my-trainings/:username", async (req, res) => {
+  const { username } = req.params;
+
+  if (!username) throw new Error("Invalid username");
+
   try {
-    const trainings = await Training.find({ trainer: "Test trainer name" });
+    const trainings = await Training.find({ trainer: username });
     res.status(200).send(trainings);
   } catch (error) {
     res.status(500).send(error);
@@ -23,6 +32,8 @@ router.get("/my-trainings", async (req, res) => {
 
 router.get("/:id", async (req, res) => {
   const { id } = req.params;
+
+  if (!id || id === "") throw new Error("Invalid ID");
 
   try {
     const training = await Training.findById(id);
@@ -36,6 +47,9 @@ router.post("/", async (req, res) => {
   const training = req.body;
 
   try {
+    isTrainingValid(training);
+    isStartAndEndDateValid(training.startDate, training.endDate);
+    isStartAndEndTimeValid(training.startTime, training.endTime);
     await Training.create(training);
     res.status(201).send(training);
   } catch (error) {
@@ -48,7 +62,12 @@ router.put("/:id", async (req, res) => {
   const { id } = req.params;
   const training = req.body;
 
+  if (!id || id === "") throw new Error("Invalid ID");
+
   try {
+    isTrainingValid(training);
+    isStartAndEndDateValid(training.startDate, training.endDate);
+    isStartAndEndTimeValid(training.startTime, training.endTime);
     const updatedTraining = await Training.findByIdAndUpdate(id, training);
     res.status(200).send(updatedTraining);
   } catch (error) {
@@ -59,6 +78,8 @@ router.put("/:id", async (req, res) => {
 
 router.delete("/:id", async (req, res) => {
   const { id } = req.params;
+
+  if (!id || id === "") throw new Error("Invalid ID");
 
   try {
     const training = await Training.findByIdAndDelete(id);
